Optionally parallelize via DDP
This adds parallelization support to the whole code base with the
assumption that the non-cINN models do not have any I/O except that I/O
which is called from a `Trainer` instance.

Parallelization is completely optional and should work with both CPU and
GPU usage.

We average any metrics such as loss to ease analysis across
processes (only need to look at statistics from one process).

Additionally, we supply a Slurm script to start a distributed training
run. This expects (1) a `pic_env_dir` and (2) a `venv_dir`
containing (1) PIConGPU libraries and according path modifications
and (2) a Python virtual environment.

Data loaders need a `DistributedSampler` in order to shard the data set.
This is supplied everywhere a `DataLoader` was encountered. Shuffling
was preserved as is, although the order after shuffling changed due to
`DistributedSampler` using a `seed=0` by default.